created super class for action and event
